Fix animation save and show.
Fixed a bug where a call to Animation.animate with a filename and show
set to True did not produce the same result.

The solution was to run the maze_walker and save the position of each
walker into two arrays (_xs and _ys) and then use that array for the
animation itself.